+++ app/(header-home)/layout.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next';
 import { NPSBanner } from '#components/banner/nps';
 import Footer from '#components/footer';
-import { HeaderServer } from '#components/header/header-server';
+import { HeaderAppRouter } from '#components/header/header-app-router';
 import { meta } from '#components/meta/meta-server';
 import SocialNetworks from '#components/social-network';
 import getSession from '#utils/server-side-helper/app/get-session';
@@ -18,7 +18,7 @@ export default async function HomeLayout({
   return (
     <>
       <NPSBanner />
-      <HeaderServer useSearchBar={false} useAgentCTA={true} />
+      <HeaderAppRouter useSearchBar={false} useAgentCTA={true} />
       <main className="fr-container">{children}</main>
       <SocialNetworks />
       <QuestionOrFeedback session={session} />
