+++ app/not-found.tsx
@@ -1,12 +1,12 @@
 import { ErrorNotFoundExplanations } from '#components/error-explanations';
 import Footer from '#components/footer';
-import { HeaderServer } from '#components/header/header-server';
+import { HeaderAppRouter } from '#components/header/header-app-router';
 import QuestionOrFeedback from './_component/question-or-feedback';
 
 export default function NotFound() {
   return (
     <>
-      <HeaderServer useSearchBar={true} useAgentCTA={true} />
+      <HeaderAppRouter useSearchBar={true} useAgentCTA={true} />
       <main className="fr-container">
         <ErrorNotFoundExplanations />
       </main>
